Update lib/LWP/Simple.pm
Had an IRC chat with moritz.  He thought my coding wasn't so bad but didn't fit with the surrounding code and suggested this instead.  During the discussion it also came up that recv doesn't support binary very well yet but $socket.read(nnnn) does.  It occurs to me that since HTTP headers often have a Content-length the read solution may be good in the nearish future.